skills/websearch/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.4.1
